sailui: break core-to-UI edges ahead of package split
