+++ coinshift/test/providers/analytics_provider_test.dart
@@ -3,6 +3,7 @@ import 'package:coinshift/providers/swap_provider.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:get_it/get_it.dart';
 import 'package:logger/logger.dart';
+import 'package:sail_ui/mocks/mocks.dart';
 import 'package:sail_ui/sail_ui.dart';
 
 void main() {
