+++ sail_ui/lib/rpcs/coinshift_rpc.dart
@@ -16,7 +16,7 @@ import 'package:sail_ui/gen/coinshift/v1/coinshift.pb.dart' as pb;
 import 'package:sail_ui/rpcs/orchestrator_rpc.dart';
 import 'package:sail_ui/rpcs/rpc_sidechain.dart';
 import 'package:sail_ui/rpcs/thunder_utxo.dart';
-import 'package:sail_ui/widgets/components/core_transaction.dart';
+import 'package:sail_ui/models/core_transaction.dart';
 
 /// API to the CoinShift server.
 abstract class CoinShiftRPC extends SidechainRPC {
