+++ sail_ui/lib/rpcs/truthcoin_rpc.dart
@@ -15,7 +15,7 @@ import 'package:sail_ui/gen/truthcoin/v1/truthcoin.pb.dart' as pb;
 import 'package:sail_ui/rpcs/orchestrator_rpc.dart';
 import 'package:sail_ui/rpcs/rpc_sidechain.dart';
 import 'package:sail_ui/rpcs/thunder_utxo.dart';
-import 'package:sail_ui/widgets/components/core_transaction.dart';
+import 'package:sail_ui/models/core_transaction.dart';
 
 /// API to the truthcoin server.
 abstract class TruthcoinRPC extends SidechainRPC {
