+++ sail_ui/lib/sail_ui.dart
@@ -132,7 +132,6 @@ export 'gen/utils/v1/utils.pbenum.dart';
 export 'gen/utils/v1/utils.pbjson.dart';
 export 'gen/utils/v1/utils.pbserver.dart';
 export 'logger.dart';
-export 'mocks/mocks.dart';
 export 'pages/create_wallet_page.dart';
 export 'pages/error_screen.dart';
 export 'pages/log_page.dart';
@@ -178,7 +177,7 @@ export 'providers/sidechain/sidechain_transactions_provider.dart';
 export 'providers/download_provider.dart';
 export 'providers/sync_provider.dart';
 export 'providers/update_provider.dart';
-export 'providers/homepage_provider.dart';
+export 'widgets/homepage/homepage_provider.dart';
 export 'providers/log_provider.dart';
 export 'gen/orchestrator/v1/bitcoin_conf.connect.client.dart';
 export 'gen/orchestrator/v1/bitcoin_conf.connect.spec.dart';
@@ -211,6 +210,7 @@ export 'rpcs/truthcoin_rpc.dart';
 export 'rpcs/photon_rpc.dart';
 export 'rpcs/zside_rpc.dart';
 export 'sail_app.dart';
+export 'window_app.dart';
 export 'services/code_search_service.dart';
 export 'utils/menu_commands.dart';
 export 'pages/settings_appearance.dart';
@@ -258,7 +258,7 @@ export 'widgets/chips/action_header_chip.dart';
 export 'widgets/chips/connection_error_chip.dart';
 export 'widgets/chips/connection_success_chip.dart';
 export 'widgets/components/bullet_point.dart';
-export 'widgets/components/core_transaction.dart';
+export 'models/core_transaction.dart';
 export 'widgets/components/mnemonic_grid.dart';
 export 'widgets/components/progress_step_tile.dart';
 export 'widgets/components/stat_row.dart';
