+++ thunder/test/ui_integration_test.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:get_it/get_it.dart';
 import 'package:logger/logger.dart';
+import 'package:sail_ui/mocks/mocks.dart';
 import 'package:sail_ui/sail_ui.dart';
 
 import 'mocks/rpc_mock_sidechain.dart';
